ledger: don't normalize types for on chain scripts
We did this for verisimilitude, but if we're going to erase on chain
then it makes no difference. Plus, this has been a bottleneck for
testing time in the past, so getting rid of it will save us time in the
short term.

The implementation continues to typecheck things (useful as a sanity
check), but the lack of normalization doesn't hurt that.